Gradlize the Memcahced-Java-Client
A lot of unit tests had failed on the Gradle. This might be caused
since all unit tests are executed in a single process on the Gradle
and the SockIOPool with the same name is shared between tests for TCP
connection and tests for UDP connection.
To pevent this issue, the following changes are added:
1) Use the different name to SockIOPool for UDP tests.
2) Check the proctrol paramter more strictly in the
   SchoonerSocketIOPool$getInstance method to detect this matter
   earlier.
3) Fetch the SocketIOPool for UDP explicitly in AscIIUDPClient's
   constructor.